refactor Mocket into separate module
